requirements.txt
discord.py==2.4.0
redis==5.0.4
aiosqlite==0.19.0
pytz==2024.1
boto3==1.34.84
cassandra-driver==3.29.0
onnxruntime==1.17.1
pillow==10.2.0
numpy==1.26.4

